+++ waku/v2/protocol/waku_store/waku_store.nim
@@ -375,7 +375,7 @@ method init*(ws: WakuStore) =
   if ws.store.isNil:
     return
 
-  proc onData(timestamp: uint64, msg: WakuMessage, pubsubTopic:  string) =
+  proc onData(timestamp: float64, msg: WakuMessage, pubsubTopic:  string) =
     # TODO index should not be recalculated
     ws.messages.add(IndexedWakuMessage(msg: msg, index: msg.computeIndex(), pubsubTopic: pubsubTopic))
